+++ packages/app/cypress/e2e/settings.cy.ts
@@ -564,7 +564,7 @@ describe('App: Settings without cloud', () => {
     })
 
     cy.contains('button', 'Connect to Cypress Cloud').click()
-    cy.findByRole('dialog', { name: 'Log in to Cypress' }).within(() => {
+    cy.findByRole('dialog', { name: 'Continue in your browser' }).within(() => {
       cy.contains('button', 'Log in').click()
     })